PROGRESS REPORT
Upon upgrading to Unity 5, there were some areas of the
source code that needed updating, due to deprecated
functions and properties.
Much of the progress is related to the visual aspect
of the game. The example room is fully textured, the
textures used may not be the final textures, however
the application, and development of the textures, their
normal maps, bump maps and speculars are merely a means
of gaining experience within that realm.
After some tinkering with Blender, I have created several meshes, static and dynamic, of which includes the ramp, we have required for some time now.
A couple of bugs have also been fixed.

Summarised Modifications:

- Ramps
- Static Meshes (Decor)
- Textures
- Push Block bounds fixed
- Floor collision audio fixed (no floor collisions)
- Deprecated source modifications

TODO:

Create more puzzles rooms, interactive objects and apply logic where needed.
Once we have enough puzzle rooms, we can install them
into the switching alglorithm which will then require a few minor adjustments.