UPDATES
I have found a lovely set of glass doors with a surround we can colour. The doors open outwards
and this achieves a far greater effect than the previous door mechanism.
I have implemented a Spawn Point custom component
which does what it says on the tin really, although
it can be useful for setting nodes for other objects
also. The Map class I wrote is implemented in the
Spawn Class, as I needed a Key/Value pair for Spawn
locations, a plain old array or list just won't do.

SUMMARY

- More optimizations and tweaks
- Added custom door mesh, testing in progress
- Added ReadOnly class, for Inspector properties
- Added Map class (inherits from Dictionary)
- Added Spawn Point component class