UPDATES
- Last level on current layer done: YAYYYY!!!
- Code optimizations

TODO:

We, the team must decide on the road ahead as there may
be conflicting code, due to different authors and we
all have our own ideas naturally. We must put pride
aside and commence with the best possible combination
of our efforts.
We also need to get all these levels into the switch
algorithm, although that goes without saying.
We need to put our great minds together to come up with
a nifty solution to fast as fuck graphics.
Other than that, it's all good (in the hood).
;)